// src/model/book.ts
type Book = {
  posterUrl: string;
  author: string;
  numberOfPages: number;
  price: number;
  yearOfPublication: string;
  isbn: string;
  publishingHouse: string;
  language: string;
  title: string;
  rating: number;
  description: string;
  letter: string;
  coverImage: string;
  bindingType: string;
  id: string;
};

export default Book;